move utils from org.kafka... => to org.kafka.storage.internals to prevent checkstyle errors
